More fixes

+++ Wflow/src/bmi.jl
@@ -3,7 +3,7 @@
 
 # Mapping of grid identifier to a key, to get the active indices of the model domain.
 # See also function active_indices(network, key::AbstractString).
-const GRIDS = Dict{Int, String}(
+const GRIDS = Dict{Int,String}(
     0 => "reservoir",
     1 => "drain",
     2 => "river",
@@ -75,7 +75,7 @@ function BMI.finalize(model::Model)
         write_netcdf_timestep(model, writer.state_dataset, writer.state_parameters)
     end
     reset_clock!(model.clock, config)
-    close_files(model; delete_output = false)
+    close_files(model; delete_output=false)
     return nothing
 end
 
@@ -153,7 +153,7 @@ end
 function BMI.get_var_units(model::Model, name::String)
     (; land) = model
     metadata = get_metadata(name, land)
-    return to_string(to_SI(metadata.unit); BMI_standard = true)
+    return to_string(to_SI(metadata.unit); BMI_standard=true)
 end
 
 function BMI.get_var_itemsize(model::Model, name::String)
@@ -176,7 +176,7 @@ function BMI.get_current_time(model::Model)
     (; starttime, calendar) = config.time
     starttime = cftime(starttime, calendar)
     # ms => s
-    return to_SI(Dates.value(clock.time - starttime), MS)
+    return to_SI(Float64(Dates.value(clock.time - starttime)), MS)
 end
 
 function BMI.get_start_time(::Model)
@@ -328,21 +328,21 @@ function BMI.get_grid_edge_nodes(model::Model, grid::Int, edge_nodes::Vector{Int
     # inactive nodes (boundary/ghost points) are set at -999
     if grid == 3
         nodes_at_edge = adjacent_nodes_at_edge(domain.river.network.graph)
-        nodes_at_edge.dst[nodes_at_edge.dst .== m + 1] .= -999
-        edge_nodes[range(1, n; step = 2)] = nodes_at_edge.src
-        edge_nodes[range(2, n; step = 2)] = nodes_at_edge.dst
+        nodes_at_edge.dst[nodes_at_edge.dst.==m+1] .= -999
+        edge_nodes[range(1, n; step=2)] = nodes_at_edge.src
+        edge_nodes[range(2, n; step=2)] = nodes_at_edge.dst
         return edge_nodes
     elseif grid == 4
         xu = domain.land.network.edge_indices.xu
-        edge_nodes[range(1, n; step = 2)] = 1:m
-        xu[xu .== m + 1] .= -999
-        edge_nodes[range(2, n; step = 2)] = xu
+        edge_nodes[range(1, n; step=2)] = 1:m
+        xu[xu.==m+1] .= -999
+        edge_nodes[range(2, n; step=2)] = xu
         return edge_nodes
     elseif grid == 5
         yu = domain.land.network.edge_indices.yu
-        edge_nodes[range(1, n; step = 2)] = 1:m
-        yu[yu .== m + 1] .= -999
-        edge_nodes[range(2, n; step = 2)] = yu
+        edge_nodes[range(1, n; step=2)] = 1:m
+        yu[yu.==m+1] .= -999
+        edge_nodes[range(2, n; step=2)] = yu
         return edge_nodes
     elseif grid in 0:2 || grid == 6
         @warn("edges are not provided for grid type $grid (variables are located at nodes)")
@@ -406,7 +406,7 @@ Return the grid element type of a model variable (PropertyLens `var`) based on a
 function grid_element_type(
     ::T,
     var::PropertyLens,
-) where {T <: Union{LocalInertialRiverFlow, LocalInertialOverlandFlow}}
+) where {T<:Union{LocalInertialRiverFlow,LocalInertialOverlandFlow}}
     vars = (PropertyLens(x) for x in (:q, :q_av, :qx, :qy))
     element_type = if var in vars
         "edge"
